+++ src/mediarithmics/plugins/audience-feed-connector/AudienceFeedConnectorBasePlugin.ts
@@ -311,6 +311,9 @@ export abstract class AudienceFeedConnectorBasePlugin extends BasePlugin<Audienc
             case 'error':
               statusCode = 500;
               break;
+            case 'retry':
+              statusCode = 429;
+              break;
             default:
               statusCode = 500;
           }