+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# xxxxx - YYYY-MM-DD
+- Enable to return retry statusName (429 Status Code) for the onUserSegmentUpdate method for the Audience Feed Connector Plugin
+
 # 0.9.1 - 2022-01-14
 
 - Fix packaging issue (size)